feat(directory_monitor): added directory_monitor
This module is a prerequisite to the hot swap functionality.
It scans a specified file on a continuous loop to see if any files
have been added, deleted, or modified since the last scan.

An event based system then triggers if changes found. Hot swap (or
other functionality) can subscribe to these events.

This additionally creates a new 'src' folder for migrating source code.