feat(bundler): add module that bundles source files into a zip archive
Blender requires add-ons with more than one file to be in a .zip archive
with a specific file structure. This commit is not the final one for this
module, but it is at a point where it will provide the required effect as
long as the user makes sure the input source list is appropriate for
Blender. TODO: Add more automated safety checks.