build: add a project build script and update bundler
In order for Blender to load the add-on, the bundled files all have to be in a single folder within a .zip
archive. On installation, Blender extracts the folder to a place such as:
    "C:\Users\[name]\AppData\Roaming\Blender Foundation\Blender\3.3\scripts\addons"
Thus, the folder "[safe_name]" will get added to this add-on location. Otherwise, it just dumps all of the
individual files in there and Blender can't figure out what to do with it.