test: restore manual debugpy test code
I have not found a way to automate this yet. This script starts a Debugpy
server that the user can manually connect to with VS Code to verify that
Python and Debugpy are working correctly. If the test does not throw an
error and the user can't connect, then the problem is with VS Code.