test(directory_monitor): improve test stability
There are several tests that will ocassionaly unexplicably fail. This is
very frustrating because running it immediately again will pass. I am not
sure what the source of that bug is, but this change improves several that
were habitual troublemakers to give the system a chance to detect the
changes it should detect and then catch up.