feat: connect the directory monitor to the debugger via the UI
This commit attaches the monitor to the debugger. It does not yet take any
actions on changes, but changes are definitely detected and printed to the
console. Console message output was improved based on formatting issues
I noted in Blender's terminal where it didn't recognize the new line escape
character (backslash + n).