feat: add hot swap capability
This is a significant content update that connects the directory monitor
with a hot swap script. The hot swap subscribes a reload function to the
monitor that takes the monitored addon and uses Blender to disable,
reinstall, and reenable it. This bypasses the need to manually install
and reload, and most importantly prevents you from having to shut down
Blender and restart the entire program.

TODO: More error checking on the reload_modules function to verify the
scripts meet the bare minimum requirements to load in Blender, as well as
connecting the bundler to the reload automation.