docs(readme): make initial updates for Blender Scripting Assistant
This is a significant content update to the README. Following the fork
from Alan North's "Blender Debugger for VS Code v2.2.0", I changed a
significant amount of functionality. It no longer follows the same scope
that Alan had, and thus the README needs significant revision. This commit
is the first of many to get the documentation ready for other developers.

Resolves #2