fix: correct bug that prevented hot swapping complex add-on packages
The `create_addon_name` function was significantly reworked. It now only
formats a provided name into the required format specification. The
accompanying tests for this were updated as well.

The logic to get the most up-to-date `bl_info.name` value is now in a
stand-alone function: `get_most_recent_bl_name_info`. This no longer uses
`__import__` and `importlib.refresh()`. Instead, it uses the `importlib`
utilities package to get a module spec based on a file location, converts
that into a module, and tries to load it. New tests were written to verify
functionality.

Finally, other logic was added to prevent Blender from crashing by trying
to remove the Blender Scripting Assistant add-on. This became necessary
while fixing this bug, so it was not split into a separate bug report.

This resolves #11.